refactor: gate simd `norm_l{1,2}` behind `simd` feature
simd is now opt-in so that library users can stay on the stable rust
toolchain
